+++ src/courseware/course/content-tools/ContentTools.jsx
@@ -21,7 +21,6 @@ export default function ContentTools({
   );
 }
 
-
 ContentTools.propTypes = {
   course: PropTypes.shape({
     notes: PropTypes.shape({
