+++ src/courseware/course/content-tools/calculator/Calculator.jsx
@@ -388,5 +388,4 @@ Calculator.propTypes = {
   intl: intlShape.isRequired,
 };
 
-
 export default injectIntl(Calculator);
