+++ src/courseware/data/index.js
@@ -1,0 +1,14 @@
+export {
+  fetchCourse,
+  fetchSequence,
+  checkBlockCompletion,
+  saveSequencePosition,
+} from './thunks';
+export {
+  getResumeBlock,
+} from './api';
+export {
+  sequenceIdsSelector,
+  firstSequenceIdSelector,
+} from './selectors';
+export { reducer } from './slice';
