--- src/data/index.js
@@ -1,7 +1,0 @@
-export {
-  fetchCourse,
-  fetchSequence,
-} from './thunks';
-
-export { getResumeBlock, normalizeBlocks } from './api';
-export { reducer } from './slice';
